requirements.txt
Django==4.2.1
ShopifyAPI==12.3.0
python-dotenv==1.0.0
django-extensions==3.2.1
Werkzeug==2.3.4
pyOpenSSL==23.1.1
honcho==1.1.0
whitenoise==6.4.0
openai==0.27.7
